SFM에 hitBox EnterCollider 구현
SFM에 히트박스부분에 부딪히면 Status.stun으로 상태를 변화시키는 부분 추가

그리고 유전자알고리즘과 팔돌리기 스크립트를 하나로 합침
