--- Ai_Project_Team_4/Assets/_Scripts/System/WeightGeneticSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0d9d645680b7d8b4082b474ea7fb3e92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
